requirements.txt
torchvision
pillow
matplotlib
opencv-python
pandas